Post 관련 문제점 해결

+++ server/src/post/post.service.ts
@@ -1,4 +1,9 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
+import { UserService } from 'src/user/user.service';
 import { Post, PostId } from './domain/post';
 import { CreatePostDto } from './dto/create-post.dto';
 import { ModifyPostDto } from './dto/modify-post.dto';
@@ -6,10 +11,18 @@ import { PostRepository } from './persistence/post.repository';
 
 @Injectable()
 export class PostService {
-  constructor(private readonly postRepository: PostRepository) {}
+  constructor(
+    private readonly userService: UserService,
+    private readonly postRepository: PostRepository,
+  ) {}
 
-  createPost(dto: CreatePostDto): Promise<Post> {
-    const post = Post.builder().set('content', dto.content).build();
+  async createPost(username: string, dto: CreatePostDto): Promise<Post> {
+    const user = await this.userService.getUserByUsername(username);
+
+    const post = Post.builder()
+      .set('user', user)
+      .set('content', dto.content)
+      .build();
     return this.postRepository.save(post);
   }
 
@@ -23,15 +36,25 @@ export class PostService {
     return post;
   }
 
-  async modifyPost(postId: PostId, dto: ModifyPostDto): Promise<Post> {
+  async modifyPost(
+    postId: PostId,
+    username: string,
+    dto: ModifyPostDto,
+  ): Promise<Post> {
     const post = await this.getPost(postId);
+
+    if (username !== post.user.username)
+      throw new ForbiddenException('작성자만 수정 가능합니다.');
 
     dto.content && post.changeContent(dto.content);
     return this.postRepository.save(post);
   }
 
-  async deletePost(postId: PostId): Promise<void> {
+  async deletePost(postId: PostId, username: string): Promise<void> {
     const post = await this.getPost(postId);
+
+    if (username !== post.user.username)
+      throw new ForbiddenException('작성자만 삭제 가능합니다.');
 
     await this.postRepository.delete(post.id);
   }
